Refactor marketplace code into separate module
- Move marketplace models to main/marketplace/models.py
- Move marketplace serializers to main/marketplace/serializers.py
- Move marketplace views to main/marketplace/views.py
- Move marketplace admin to main/marketplace/admin.py
- Translate all Russian comments to English
- Reduce main app file sizes by ~600 lines

This refactoring improves code organization and maintainability
by separating marketplace-specific code from the main app while
maintaining backward compatibility through imports.
