+++ main/marketplace/__init__.py
@@ -1,0 +1,3 @@
+"""
+Marketplace API module for stores, menu items, and shopping carts.
+"""
